First working MQTT version

+++ umqtt/umqtt.c
@@ -31,10 +31,10 @@
 #define umqtt_header_type(h) \
     ((h) >> 4)
 
-static int umqtt_decode_length(uint8_t *data) {
-    int mul = 1;
-    int val = 0;
-    int i;
+static int16_t umqtt_decode_length(uint8_t *data) {
+    int16_t mul = 1;
+    int16_t val = 0;
+    int16_t i;
 
     for (i = 0; i == 0 || (data[i - 1] & 0x80); i++) {
         val += (data[i] & 0x7f) * mul;
@@ -43,9 +43,9 @@ static int umqtt_decode_length(uint8_t *data) {
     return val;
 }
 
-static int umqtt_encode_length(int len, uint8_t *data) {
-    int digit;
-    int i = 0;
+static int16_t umqtt_encode_length(int16_t len, uint8_t *data) {
+    int16_t digit;
+    int16_t i = 0;
 
     do {
         digit = len % 128;
@@ -63,7 +63,7 @@ void umqtt_circ_init(struct umqtt_circ_buffer *buff) {
     buff->datalen = 0;
 }
 
-int umqtt_circ_push(struct umqtt_circ_buffer *buff, uint8_t *data, int len) {
+int16_t umqtt_circ_push(struct umqtt_circ_buffer *buff, uint8_t *data, int16_t len) {
     uint8_t *bend = buff->start + buff->length - 1;
     /* This points to new byte */
     uint8_t *dend = (buff->pointer - buff->start + buff->datalen) % buff->length + buff->start;
@@ -81,10 +81,10 @@ int umqtt_circ_push(struct umqtt_circ_buffer *buff, uint8_t *data, int len) {
     return len; /* Return amount of bytes left */
 }
 
-int umqtt_circ_peek(struct umqtt_circ_buffer *buff, uint8_t *data, int len) {
+int16_t umqtt_circ_peek(struct umqtt_circ_buffer *buff, uint8_t *data, int16_t len) {
     uint8_t *ptr = buff->pointer;
     uint8_t *bend = buff->start + buff->length - 1;
-    int i;
+    int16_t i;
 
     for (i = 0; i < len && i < buff->datalen; i++) {
         data[i] = ptr[i];
@@ -94,9 +94,9 @@ int umqtt_circ_peek(struct umqtt_circ_buffer *buff, uint8_t *data, int len) {
     return i; /* Return the amount of bytes actually peeked */
 }
 
-int umqtt_circ_pop(struct umqtt_circ_buffer *buff, uint8_t *data, int len) {
+int16_t umqtt_circ_pop(struct umqtt_circ_buffer *buff, uint8_t *data, int16_t len) {
     uint8_t *bend = buff->start + buff->length - 1;
-    int i;
+    int16_t i;
 
     for (i = 0; i < len && buff->datalen > 0; i++) {
         data[i] = *buff->pointer;
@@ -117,7 +117,7 @@ void umqtt_init(struct umqtt_connection *conn) {
 }
 
 void umqtt_connect(struct umqtt_connection *conn, uint16_t kalive, char *cid) {
-    int cidlen = strlen(cid);
+    int16_t cidlen = strlen(cid);
     uint8_t fixed;
     uint8_t remlen[4];
     uint8_t variable[12];
@@ -151,7 +151,7 @@ void umqtt_connect(struct umqtt_connection *conn, uint16_t kalive, char *cid) {
 }
 
 void umqtt_subscribe(struct umqtt_connection *conn, char *topic) {
-    int topiclen = strlen(topic);
+    int16_t topiclen = strlen(topic);
     uint8_t fixed;
     uint8_t remlen[4];
     uint8_t messageid[2];
@@ -173,8 +173,8 @@ void umqtt_subscribe(struct umqtt_connection *conn, char *topic) {
     conn->nack_subscribe++;
 }
 
-void umqtt_publish(struct umqtt_connection *conn, char *topic, uint8_t *data, int datalen) {
-    int toplen = strlen(topic);
+void umqtt_publish(struct umqtt_connection *conn, char *topic, uint8_t *data, int16_t datalen) {
+    int16_t toplen = strlen(topic);
     uint8_t fixed;
     uint8_t remlen[4];
     uint8_t len[2];
@@ -199,7 +199,7 @@ void umqtt_ping(struct umqtt_connection *conn) {
     conn->nack_ping++;
 }
 
-static void umqtt_handle_publish(struct umqtt_connection *conn, uint8_t *data, int len) {
+static void umqtt_handle_publish(struct umqtt_connection *conn, uint8_t *data, int16_t len) {
     uint16_t toplen = (data[0] << 8) | data[1];
     char topic[toplen + 1];
     uint8_t payload[len - 2 - toplen];
@@ -211,7 +211,7 @@ static void umqtt_handle_publish(struct umqtt_connection *conn, uint8_t *data, i
     conn->message_callback(conn, topic, payload, sizeof(payload));
 }
 
-static void umqtt_packet_arrived(struct umqtt_connection *conn, uint8_t header, int len) {
+static void umqtt_packet_arrived(struct umqtt_connection *conn, uint8_t header, int16_t len) {
     uint8_t data[len];
 
     umqtt_circ_pop(&conn->rxbuff, data, len);
@@ -236,7 +236,7 @@ static void umqtt_packet_arrived(struct umqtt_connection *conn, uint8_t header, 
 
 void umqtt_process(struct umqtt_connection *conn) {
     uint8_t buf[5];
-    unsigned int i = 2;
+    uint16_t i = 2;
 
     while (conn->rxbuff.datalen >= 2) { /* We do have the fixed header */
         umqtt_circ_pop(&conn->rxbuff, buf, 2);
